[IE][GNA][TESTS] Move ImportExport tests from GNA to shared part
