+++ inference-engine/tests/functional/plugin/shared/include/import_export_tests/import_reshape_permute_conv.hpp
@@ -1,0 +1,16 @@
+// Copyright (C) 2020 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include "import_export_base.hpp"
+
+namespace LayerTestsDefinitions {
+
+class ImportReshapePermuteConv : public ImportNetworkTestBase {
+protected:
+    void SetUp() override;
+};
+
+} // namespace LayerTestsDefinitions
